+++ eveuniverse/tests/test_core.py
@@ -8,7 +8,7 @@ from django.core.cache import cache
 from django.test import TestCase
 
 from ..constants import EveGroupId
-from ..core import esitools, eveimageserver, evemicros, eveskinserver
+from ..core import dotlan, esitools, eveimageserver, evemicros, eveskinserver, evewho
 from ..utils import NoSocketsTestCase
 from .testdata.esi import EsiClientStub
 from .testdata.factories import create_evemicros_request
@@ -295,3 +295,54 @@ class TestEveMicrosNearestCelestial(TestCase):
         )
         # then
         self.assertEqual(result.id, 40170699)
+
+
+class TestEveWho(TestCase):
+    def test_alliance_url(self):
+        self.assertEqual(
+            evewho.alliance_url(12345678), "https://evewho.com/alliance/12345678"
+        )
+
+    def test_corporation_url(self):
+        self.assertEqual(
+            evewho.corporation_url(12345678), "https://evewho.com/corporation/12345678"
+        )
+
+    def test_character_url(self):
+        self.assertEqual(
+            evewho.character_url(12345678), "https://evewho.com/character/12345678"
+        )
+
+
+class TestDotlan(TestCase):
+    def test_alliance_url(self):
+        self.assertEqual(
+            dotlan.alliance_url("Wayne Enterprices"),
+            "https://evemaps.dotlan.net/alliance/Wayne_Enterprices",
+        )
+
+    def test_corporation_url(self):
+        self.assertEqual(
+            dotlan.corporation_url("Wayne Technology"),
+            "https://evemaps.dotlan.net/corp/Wayne_Technology",
+        )
+        self.assertEqual(
+            dotlan.corporation_url("Crédit Agricole"),
+            "https://evemaps.dotlan.net/corp/Cr%C3%A9dit_Agricole",
+        )
+
+    def test_faction_url(self):
+        self.assertEqual(
+            dotlan.faction_url("Amarr Empire"),
+            "https://evemaps.dotlan.net/factionwarfare/Amarr_Empire",
+        )
+
+    def test_region_url(self):
+        self.assertEqual(
+            dotlan.region_url("Black Rise"), "https://evemaps.dotlan.net/map/Black_Rise"
+        )
+
+    def test_solar_system_url(self):
+        self.assertEqual(
+            dotlan.solar_system_url("Jita"), "https://evemaps.dotlan.net/system/Jita"
+        )