+++ CHANGELOG.md
@@ -9,7 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- EveEntity.profile_url: URL to display this type on the default third party webpage
 - EveType.profile_url: URL to display this type on the default third party webpage
+- core.evewho: URL for profile pages on evewho webpage
+- core.dotlan: URL for profile pages on dotlan webpage
 
 ## [0.9.0] - 2021-12-30